+++ backend/packages/provider-s3/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @cat-factory/provider-s3
 
+## 0.2.282
+
+### Patch Changes
+
+- Updated dependencies [36b1853]
+  - @cat-factory/kernel@0.225.0
+
+## 0.2.281
+
+### Patch Changes
+
+- Updated dependencies [413095f]
+  - @cat-factory/kernel@0.224.0
+
 ## 0.2.280
 
 ### Patch Changes